Google maps fully integrated
Once when the user dropped a image or selected a Image it will
automatically prompt user to select location from the map. In the map a
rectangle is drawn to denote the hunt area and user can select a
location just clicking inside the hunt area. Once the user selected a
location he can click again on the marker to return to the form.